+++ RELEASES.md
@@ -8,6 +8,7 @@ Released on TBD (UTC).
 - Added `TardisInstrumentProvider`
 - Added `PolymarketExecClientConfig.generate_order_history_from_trades` config setting (default False and not currently recommended)
 - Added configuration warning for `InstrumentProvider` (to warn when node starts with no instrument loading)
+- Implemented Tardis optional [symbol normalization](https://nautilustrader.io/docs/nightly/integrations/tardis/#symbology-and-normalization)
 - Implemented `WebSocketClient` reconnection retries (#2044), thanks @davidsblom
 - Implemented `OrderCancelRejected` event generation for Binance and Bybit
 - Implemented `OrderModifyRejected` event generation for Binance and Bybit